chore: clean up base class and export types
