+++ 2nd-gen/packages/core/components/action-button/index.ts
@@ -9,5 +9,4 @@
  * OF ANY KIND, either express or implied. See the License for the specific language
  * governing permissions and limitations under the License.
  */
-export * from './ActionButton.base.js';
 export * from './ActionButton.types.js';
